Use Context pervasively
It is recommended to set timeouts on all outgoing requests in order
to prevent indefinitely open connections or other dangling resources.
The way to do that is with the stdlib Context, which conventionally
occupies the first function parameter.

This change uses Context in two ways:

1. Propagating Context of an incoming HTTP request in the handler
   (this is provided by the Go HTTP server per request).
2. Adding ctx params to the Storage interface methods and passing
   inherited Contexts to them.

Observing Contexts provided to handlers also means we can abort work
early if the client has canceled the request.

As the http package was stabilized before Context was introduced,
most HTTP calls have to be made with http.NewRequestWithContext
instead of the convenient http.Get, http.PostForm, etc.

The asynchronous operations are given Contexts detached from the
original parent; otherwise, the request would end and the workload
would be canceled.

I added some timeout fields to Config with default values (we can
change these durations as appropriate for the workloads). As an
oversight in the Go standard library, time.Duration does not natively
support JSON marhaling/unmarhaling, so I had to make a small wrapper
type for these.